+++ server/api/multipleChoice/[id].get.js
@@ -1,0 +1,13 @@
+import { PrismaClient } from "@prisma/client";
+const prisma = new PrismaClient();
+
+export default defineEventHandler(async (event) => {
+  // return event.context.params.id;
+  return await prisma.multipleChoice.findFirst({
+    where: { id: parseInt(event.context.params.id) },
+    select: {
+      title: true,
+      question: true
+    }
+  });
+});